Make sure all hidden files are outside the background image
If the user enabled Mac OS X to show hidden files then when mounted the
generated DMG package may show some of them within the bounds of the
background image.

This change makes sure that all items in the DMG package are moved to a
position outside the bounds of the background image. The
POSITITION_CLAUSE and APPLICATION_CLAUSE that are executed later will
still place the app icon and Application correctly.