add option to set custom Rez path
On some systems the Rez tool isn't installed on the default
/Applications/Xcode.app/Contents/Developer/Tools/Rez

Currently when using the --eula flag in order to embed a license in
the DMG file, the process can fail when the dmg-license.py script
is called if Rez can't be found in the default location.

This patch adds a --rez flag which allows users to specify a custom
Rez path which is then passed to dmg-license.py, as this script
already has support for using a custom Rez path.