Make sure target image size is within limits.
Each image has a minimum size based on how it was formatted. If the target size
we ask for is less than that, hdiutil will fail with `hdiutil: resize: failed.
Invalid argument (22)`. To avoid this, find the minimum size with `hdiutil
resize -limits` and use that if it's larger than the target size.