Enable multiple parallel calls
In a CI/CD environment, it is possible that multiple jobs are running
in parallel. In such case, the script will fail because the dmg file is
already created by another job.

This commit adds a the PID of the process creating the dmg file to the
name of the temporary dmg file. It also uses -mountrandom to mount the
dmg file in a random directory. This way, multiple jobs can run in
parallel.

Signed-off-by: Mikaël Barbero <mikael.barbero@eclipse-foundation.org>