Added option to leave rw image
This PR adds a check to see if the target file already exists, as it otherwise errors out when creating the final output.

To fix this there is now a `--overwrite` option that enables overwriting the output file.

To support cases where further customization of the image is required, there is also a `--skip-finalize` option that returns the intermediate write-enabled dmg file.